new classes for // home page, profile, all scenario screens, trade //
Notes for tab method:::
Parameter method COULD work
If what we do is:
for each scenario, case, whatever we feed different values to the screen, so that for each case the screen would acc be unique and specific. and dont have to worry about the other cases until they are referenced. with this being said all the other possible cases and their values do not acc matter until its their turn to show up.